Compatibility with clang, for supporting rust-bindgen

+++ teensy3/WProgram.h
@@ -8,8 +8,8 @@
 // some libraries and sketches depend on this
 // AVR stuff, assuming Arduino.h or WProgram.h
 // automatically includes it...
-#include <avr/pgmspace.h>
-#include <avr/interrupt.h>
+#include "avr/pgmspace.h"
+#include "avr/interrupt.h"
 
 #include "avr_functions.h"
 #include "wiring.h"
@@ -50,7 +50,7 @@ void tone(uint8_t pin, uint16_t frequency, uint32_t duration = 0);
 void noTone(uint8_t pin);
 
 // WMath prototypes
-int32_t random(void);
+long int random(void);
 uint32_t random(uint32_t howbig);
 int32_t random(int32_t howsmall, int32_t howbig);
 void randomSeed(uint32_t newseed);
